Let's use NickServ to authenticate a user.
The admin plugin has been reconfigured to use NickServ for user authentication,
and uses a configurable list in cinchize.yml to authorize administrators. This
pushes the messy part off onto the IRC network services, and allows us to just
deal with what we are really interested in. Users can now control the bot from
within IRC without needing to go to an external location to do the basic
management. In the future, we could use the same framework to allow for more
JBot management inline, without worrying about securing passwords or sessions or
whatnot else.