Allow admins to update data.json through IRC commands.
Through the !start_show <show> and !end_show commands, an announcement can be
made for a show in all channels, and the data.json file will be automatically
updated with live show information. Since it is going through the admin plugin,
users can be authenticated and authorized before being allowed to start or end a
show.

I am not totally happy with the implementation because of the varying levels of
abstraction in the methods, but it works correctly, and so is a good start.