A better way to deal with data.json.
Dealing with data.json has now been extracted into a separate model class. It
now handles all the file reading and writing internally so that we don't have to
deal with that at a higher level. It also provides attrbutes for accessing the
data bits inside of data.json without having to get messy. It does *not*
automatically rewrite the file on each attribute change. It also provides
convenience methods to handle the flow of starting and stopping a show.