Make DataJSON more robust.
This makes DataJSON more robust. The persistence parts are split out into their
own module classes. This allows the persistence backend to be pluggable. A nil
persistence is provided for when no path is provided. A default file-based class
provides access to the local file system. A new HTTP-backed persistence is
partially implemented for loading a remote file. Explicit exceptions are now
raised by the class on expected errors. These are to be handled and logged at
the plugin level, as can be seen in the admin plugin.